requirements.txt
requests==2.31.0
tiktoken==0.5.1
python-dotenv==1.0.0
openai==1.3.0
argparse==1.4.0 